+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Avoid endless retry after authentication is revoked #494
 - Avoid using cached token endpoints from OIDAuthState #487
 - macOS: Point the user to the OS' please-enable-notification prompt #474 
+- Improve error alerts #489
 
 ## 3.0.6